goalflow/analytics/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
plotly==5.18.0
loguru==0.7.2
pydantic==2.5.3
pydantic-settings==2.1.0
